+++ src/main/java/cz/cvut/fit/niadp/mvcgame/model/Position.java
@@ -1,27 +1,27 @@
 package cz.cvut.fit.niadp.mvcgame.model;
 
 public class Position {
-    private int dimX = 0;
-	private int dimY = 0;
+    private float dimX = 0;
+	private float dimY = 0;
 
-	public Position(int posX, int posY) {
+	public Position(float posX, float posY) {
 		this.dimX = posX;
 		this.dimY = posY;
 	}
 
-	public int getX() {
+	public float getX() {
 		return dimX;
 	}
     
-    public int getY() {
+    public float getY() {
 		return dimY;
 	}
     
-    public void setY(int y) {
+    public void setY(float y) {
 		this.dimY = y;
 	}
     
-    public void setX(int x) {
+    public void setX(float x) {
 		this.dimX = x;
 	}
